% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_things.R
\name{make_things}
\alias{make_things}
\title{pRocessing}
\usage{
make_things(words = "Hi!")
}
\arguments{
\item{words}{That is one argument which, at least should, be a character string.}
}
\description{
It gives R and the user possibility to contemplate and pRocess the input or a thought given.
}
